feat(discover): discover Almost Complete
Added most of the internal discover components. States for maps and the dicover page are being
updated with redux. Need to fix some functionality and restructure this. Specifically, activities
list needs to contain all filering functionality, and not have a seperate filtered list compoent. I
also dont want to render the details pages with the router. I will manage this with local states.