feat(dashboard map): added functionality
Added map functionality to dashboard but realized i need to redo a lot of work in the map reducer.
Starting from this commit, ill be splitting the activities,filtered activites, and tags states into
a discover reducer and out of the map reducer. Im going to add a page state to the map reducer
indicating if you are on discover or dashboard. This will change the interaction when clicking an
activity. LAst, I need to change the  set markers logic so it is not tied to filtered activites